Unify and simplify a lot of pixomatic code
git-svn-id: svn+ssh://svn.corp.imvu.com/home/svnrepos/trunk/stable/Source/cal3d@51002 07c76cb3-cb09-0410-85de-c24d39f1912e

+++ src/cal3d/color.h
@@ -4,7 +4,7 @@
 #include "cal3d/vector.h"
 
 
-typedef unsigned long CalColor32; // 32-bit integer, compatible with NSPR
+typedef unsigned int CalColor32; // 32-bit integer, compatible with NSPR
 
 inline CalColor32 CalMakeColor(CalVector cv) {
     return
